server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.1"
dependencies = [
    "anthropic>=0.49.0",
    "colorlog>=6.9.0",
    "daphne>=4.1.2",
    "django>=5.1.7",
    "django-cors-headers>=4.7.0",
    "django-ninja>=1.4.0",
    "django-polymorphic>=3.1.0",
    "dotenv>=0.9.9",
    "groq>=0.20.0",
    "instructor>=1.7.8",
    "langgraph>=0.3.21",
    "langgraph-api>=0.2.61",
    "langgraph-cli[inmem]>=0.3.3",
    "openai>=1.69.0",
    "psycopg>=3.2.6",
    "pydantic>=2.11.1",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "pytest-django>=4.10.0",
    "pytest-watcher>=0.4.3",
]
